added ability for user to retweet

+++ tweet.go
@@ -15,13 +15,24 @@ func tweet_create(res http.ResponseWriter, req *http.Request) {
 
 	req.ParseForm()
 	tweet_text := req.PostFormValue("tweet")
-
+	retUser := Models.User{}
 	// Get the "session_uid" cookie to get the current loged in users id
 	user_id, err := req.Cookie("session_uid")
 
-	current_time := time.Now()
+	if err != nil {
+		panic(err)
+	}
 
-	_, err = Db.Exec("insert into tweets (user_id, msg, created_at) values ($1, $2, $3)", user_id.Value, tweet_text, current_time)
+	err = Db.QueryRow("select id, name, email, username, password, created_at, updated_at from users where id = $1", user_id.Value).Scan(&retUser.Id, &retUser.Name, &retUser.Email, &retUser.Username, &retUser.Hash, &retUser.Created_at, &retUser.Updated_at)
+
+	if err != nil {
+		panic(err)
+	}
+
+	current_time := time.Now()
+	is_retweet := false
+
+	_, err = Db.Exec("insert into tweets (user_id, msg, name, username, is_retweet, origin_user_id, origin_name, origin_username, created_at) values ($1, $2, $3, $4, $5, $1, $3, $4, $6)", retUser.Id, tweet_text, retUser.Name, retUser.Username, is_retweet, current_time)
 
 	if err != nil {
 		panic(err)
@@ -46,13 +57,22 @@ func tweet_delete(res http.ResponseWriter, req *http.Request) {
 	if err == nil {
 		url_params := mux.Vars(req)
 		tweet_id := url_params["tweet_id"]
+		tweet := Models.Tweet{}
 
-		fmt.Printf("\nDeleting the tweet id of %s...", tweet_id)
-		_, err := Db.Exec("delete from tweets where id = $1", tweet_id)
-		fmt.Printf("Done.\n")
+		err = Db.QueryRow("select id, user_id, msg, name, username, is_retweet, origin_tweet_id, origin_user_id, origin_name, origin_username, created_at from tweets where id = $1", tweet_id).Scan(&tweet.Id, &tweet.User_id, &tweet.Message, &tweet.Name, &tweet.Username, &tweet.Is_retweet, &tweet.Otweet_id, &tweet.Ouser_id, &tweet.Oname, &tweet.Ousername, &tweet.Created_at)
 
 		if err != nil {
 			panic(err)
+		}
+
+		if tweet.User_id == session.User_id {
+			fmt.Printf("\nDeleting the tweet id of %s...", tweet_id)
+			_, err := Db.Exec("delete from tweets where id = $1", tweet_id)
+			fmt.Printf("Done.\n")
+
+			if err != nil {
+				panic(err)
+			}
 		}
 	}
 
@@ -63,12 +83,33 @@ func tweet_delete(res http.ResponseWriter, req *http.Request) {
 func retweet_create(res http.ResponseWriter, req *http.Request) {
 	fmt.Printf("\n\nUser accessed the '%s' url path.\n", req.URL.Path)
 
-	fmt.Printf("\nRedirecting to the '/' path\n")
-	http.Redirect(res, req, "/", 302)
-}
+	tweet := Models.Tweet{}
+	retUser := Models.User{}
+	is_retweet := true
+	current_time := time.Now()
 
-func retweet_delete(res http.ResponseWriter, req *http.Request) {
-	fmt.Printf("\n\nUser accessed the '%s' url path.\n", req.URL.Path)
+	url_params := mux.Vars(req)
+	tweet_id := url_params["tweet_id"]
+
+	user_id, err := req.Cookie("session_uid")
+
+	err = Db.QueryRow("select id, user_id, msg, name, username, is_retweet, origin_tweet_id, origin_user_id, origin_name, origin_username, created_at from tweets where id = $1 limit 1", tweet_id).Scan(&tweet.Id, &tweet.User_id, &tweet.Message, &tweet.Name, &tweet.Username, &tweet.Is_retweet, &tweet.Otweet_id, &tweet.Ouser_id, &tweet.Oname, &tweet.Ousername, &tweet.Created_at)
+
+	if err != nil {
+		panic(err)
+	}
+
+	err = Db.QueryRow("select id, name, username from users where id = $1", user_id.Value).Scan(&retUser.Id, &retUser.Name, &retUser.Username)
+
+	if err != nil {
+		panic(err)
+	}
+
+	_, err = Db.Exec("insert into tweets (user_id, msg, name, username, is_retweet, origin_tweet_id, origin_user_id, origin_name, origin_username, created_at) values ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)", retUser.Id, tweet.Message, retUser.Name, retUser.Username, is_retweet, tweet.Id, tweet.User_id, tweet.Name, tweet.Username, current_time)
+
+	if err != nil {
+		panic(err)
+	}
 
 	fmt.Printf("\nRedirecting to the '/' path\n")
 	http.Redirect(res, req, "/", 302)
